style(backend): satisfy golangci-lint (0 issues)
Bring the backend to a clean `golangci-lint run` under the strict v2
config. 121 findings -> 0.

Linter scoping (.golangci.yml)
  - wrapcheck: only require wrapping for third-party errors; errors
    crossing our own internal package boundaries (service -> repository)
    are passed through deliberately.
  - goconst: ignore gin response-envelope keys
    ("error"/"message"/"data"/"token"/"total"/"unauthorized") —
    framework idiom, not constants.

Code fixes
  - wrapcheck: wrap genuine third-party errors with %w (gorm
    Open/AutoMigrate/Transaction, redis, jwt, bcrypt, json.Marshal,
    uuid).
  - errcheck: handle previously-ignored returns (_ = for fire-and-forget
    notifications and SetTrustedProxies; deferred _ = ...Close()).
  - dupl: extract respondOwnedResourceError, deduplicating the
    post/comment update+delete handlers.
  - goconst: extract constants — controllers messages (constants.go),
    seed MIME types, oauth provider.
  - revive/govet: drop err shadowing, simplify redundant if-return,
    blank-import comment, rename genuinely-unused params to _.
  - lll: wrap long signatures and GORM query chains; extract reusable
    SQL fragments in friend_service; staticcheck S1008, unconvert,
    prealloc, andan unused mock field.

Behaviour is unchanged; full test suite still passes at ~81% coverage.
Three justified //nolint directives remain (two long gorm struct tags,
one fixed-fixture test-helper param).

+++ backend/tests/oauth_test.go
@@ -120,7 +120,7 @@ func TestOAuthService_FindOrCreateUser(t *testing.T) {
 	ctx := context.Background()
 
 	gh := &services.GitHubUser{
-		ID:        int64(time.Now().UnixNano() % 1000000000),
+		ID:        time.Now().UnixNano() % 1000000000,
 		Login:     "ghnewuser",
 		Name:      "GH New User",
 		Email:     "ghnew@example.com",
@@ -154,7 +154,7 @@ func TestOAuthService_FindOrCreateUser_LinksExistingEmail(t *testing.T) {
 	registerAndLogin(t, router, "linklocal", "linkme@example.com", "StrongPass123!")
 
 	gh := &services.GitHubUser{
-		ID:    int64(time.Now().UnixNano()%1000000000) + 1,
+		ID:    time.Now().UnixNano()%1000000000 + 1,
 		Login: "linklocal",
 		Name:  "Link Local",
 		Email: "linkme@example.com",
@@ -181,7 +181,7 @@ func TestOAuthService_FindOrCreateUser_UsernameCollision(t *testing.T) {
 	registerAndLogin(t, router, "octocat", "octocat@example.com", "StrongPass123!")
 
 	gh := &services.GitHubUser{
-		ID:    int64(time.Now().UnixNano()%1000000000) + 2,
+		ID:    time.Now().UnixNano()%1000000000 + 2,
 		Login: "octocat",
 		Email: "freshoctocat@example.com",
 	}
